Refactor FirebaseListingDataAccessObject to implement FilterByGenreDataAccessInterface and handle genre-based listing filtering
